Corrected the tests that were using the variable int EffortCost of the ProjectCollaborator Class.

+++ src/test/java/project/controller/US390CalculateReportedProjectCostControllerTest.java
@@ -184,10 +184,10 @@ public class US390CalculateReportedProjectCostControllerTest {
 		// Calculates the value of the project - Equals to to the sum of the total hours
 		// spent times the cost of each TaskWorker
 
-		int danielCost = 5 * taskWorkerDaniel.getProjectCollaboratorFromTaskCollaborator().getCollaboratorCost();
-		int jonnyCost = 10 * taskWorkerJonny.getProjectCollaboratorFromTaskCollaborator().getCollaboratorCost();
-		int mikeCost = 2 * taskWorkerMike.getProjectCollaboratorFromTaskCollaborator().getCollaboratorCost();
-		int anaCost = 3 * taskWorkerAna.getProjectCollaboratorFromTaskCollaborator().getCollaboratorCost();
+		double danielCost = 5 * taskWorkerDaniel.getProjectCollaboratorFromTaskCollaborator().getCollaboratorCost();
+		double jonnyCost = 10 * taskWorkerJonny.getProjectCollaboratorFromTaskCollaborator().getCollaboratorCost();
+		double mikeCost = 2 * taskWorkerMike.getProjectCollaboratorFromTaskCollaborator().getCollaboratorCost();
+		double anaCost = 3 * taskWorkerAna.getProjectCollaboratorFromTaskCollaborator().getCollaboratorCost();
 
 		totalCost = danielCost + jonnyCost + mikeCost + anaCost;
